Move all eval refernce to the new path
